refactor(rpc): Move verifier nonce_gen outside of rpc module.
